Init CreateGroup component stack
As a skeleton, this is a carbon copy of the Groups stack with a few variables changed.

Right now, we have an excess of stacked components and it's getting a little messy. The tentative plan is to replace the "GROUPS" nav button in Social with a dropdown after we reach MVP. That dropdown will show existing groups and have the option to bring up the CreateGroup component.